Refactor to remove duplicated code, reflect some review comments
