+++ pkg/dependency/parser/executable/nodejs/parse_test.go
@@ -1,0 +1,54 @@
+package nodejsparser
+
+import (
+	"os"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+
+	"github.com/aquasecurity/trivy/pkg/dependency/types"
+)
+
+func TestParse(t *testing.T) {
+	tests := []struct {
+		name      string
+		inputFile string
+		want      []types.Library
+		wantDep   []types.Dependency
+		wantErr   string
+	}{
+		{
+			name:      "ELF12",
+			inputFile: "testdata/node.12.elf",
+			want: []types.Library{
+				{
+					ID: 	 "node@12.16.3",
+					Name:    "node",
+					Version: "12.16.3",
+				},
+			},
+		},
+		{
+			name:      "sad path",
+			inputFile: "testdata/dummy",
+			wantErr:   "unrecognized executable format",
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			f, err := os.Open(tt.inputFile)
+			require.NoError(t, err)
+			parser := NewParser()
+			got, _, err := parser.Parse(f)
+			if tt.wantErr != "" {
+				require.NotNil(t, err)
+				require.ErrorContains(t, err, tt.wantErr)
+				return
+			}
+
+			assert.NoError(t, err)
+			assert.Equal(t, tt.want, got)
+		})
+	}
+}
