+++ pkg/dependency/parser/executable/nodejs/testdata/dummy
@@ -1,0 +1,1 @@
+AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
