+++ pkg/dependency/parser/executable/php/parse_test.go
@@ -1,0 +1,54 @@
+package phpparser
+
+import (
+	"os"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+
+	"github.com/aquasecurity/trivy/pkg/dependency/types"
+)
+
+func TestParse(t *testing.T) {
+	tests := []struct {
+		name      string
+		inputFile string
+		want      []types.Library
+		wantErr   string
+	}{
+		{
+			name:      "ELF",
+			inputFile: "testdata/php.elf",
+			want: []types.Library{
+				{
+					ID: 	 "php@8.0.7",
+					Name:    "php",
+					Version: "8.0.7",
+				},
+			},
+		},
+		{
+			name:      "sad path",
+			inputFile: "testdata/dummy",
+			wantErr:   "unrecognized executable format",
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			f, err := os.Open(tt.inputFile)
+			require.NoError(t, err)
+
+			parser := NewParser()
+			got, _, err := parser.Parse(f)
+			if tt.wantErr != "" {
+				require.NotNil(t, err)
+				require.ErrorContains(t, err, tt.wantErr)
+				return
+			}
+
+			assert.NoError(t, err)
+			assert.Equal(t, tt.want, got)
+		})
+	}
+}
