+++ pkg/dependency/parser/executable/php/testdata/dummy
@@ -1,0 +1,1 @@
+AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
