--- pkg/dependency/parser/nodejs/binary/testdata/dummy
@@ -1,1 +1,0 @@
-AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
