--- pkg/dependency/parser/python/binary/testdata/dummy
@@ -1,1 +1,0 @@
-AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
