--- pkg/fanal/analyzer/language/nodejs/binary/testdata/broken_elf
@@ -1,1 +1,0 @@
-ELF
