--- pkg/fanal/analyzer/language/nodejs/binary/testdata/symlink
@@ -1,1 +1,0 @@
-foo
