--- pkg/fanal/analyzer/language/php/binary/binary_test.go
@@ -1,107 +1,0 @@
-package generic
-
-import (
-	"context"
-	"os"
-	"runtime"
-	"testing"
-
-	"github.com/samber/lo"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-
-	"github.com/aquasecurity/trivy/pkg/fanal/analyzer"
-	"github.com/aquasecurity/trivy/pkg/fanal/types"
-)
-
-func Test_rustBinaryLibraryAnalyzer_Analyze(t *testing.T) {
-	tests := []struct {
-		name      string
-		inputFile string
-		want      *analyzer.AnalysisResult
-	}{
-		{
-			name:      "happy path",
-			inputFile: "testdata/php",
-			want: &analyzer.AnalysisResult{
-				Applications: []types.Application{
-					{
-						Type:     types.PhpGeneric,
-						FilePath: "testdata/php",
-						Libraries: types.Packages{
-							{
-								ID:        "php@8.0.7",
-								Name:      "php",
-								Version:   "8.0.7",
-							},
-						},
-					},
-				},
-			},
-		},
-		{
-			name:      "not Python binary",
-			inputFile: "testdata/dummy",
-		},
-		{
-			name:      "broken elf",
-			inputFile: "testdata/broken_elf",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			f, err := os.Open(tt.inputFile)
-			require.NoError(t, err)
-			defer f.Close()
-
-			a := phpBinaryAnalyzer{}
-			ctx := context.Background()
-			got, err := a.Analyze(ctx, analyzer.AnalysisInput{
-				FilePath: tt.inputFile,
-				Content:  f,
-			})
-
-			assert.NoError(t, err)
-			assert.Equal(t, tt.want, got)
-		})
-	}
-}
-
-
-func Test_pythonBinaryAnalyzer_Required(t *testing.T) {
-	tests := []struct {
-		name     string
-		filePath string
-		want     bool
-	}{
-		{
-			name:     "Python binary",
-			filePath: "testdata/php",
-			want:     true,
-		},
-		{
-			name:     "executable file",
-			filePath: lo.Ternary(runtime.GOOS == "windows", "testdata/binary.exe", "testdata/0755"),
-			want:     false,
-		},
-		{
-			name:     "file perm 0644",
-			filePath: "testdata/0644",
-			want:     false,
-		},
-		{
-			name:     "symlink",
-			filePath: "testdata/symlink",
-			want:     false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			a := phpBinaryAnalyzer{}
-			fileInfo, err := os.Lstat(tt.filePath)
-			require.NoError(t, err)
-			got := a.Required(tt.filePath, fileInfo)
-			assert.Equal(t, tt.want, got, fileInfo.Mode().Perm())
-		})
-	}
-}
