--- pkg/fanal/analyzer/language/php/binary/testdata/broken_elf
@@ -1,1 +1,0 @@
-ELF
