--- pkg/fanal/analyzer/language/php/binary/testdata/dummy
@@ -1,1 +1,0 @@
-AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
