--- pkg/fanal/analyzer/language/php/binary/testdata/symlink
@@ -1,1 +1,0 @@
-./foo
