--- pkg/fanal/analyzer/language/python/binary/testdata/broken_elf
@@ -1,1 +1,0 @@
-ELF
