--- pkg/fanal/analyzer/language/python/binary/testdata/dummy
@@ -1,1 +1,0 @@
-AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
