--- pkg/fanal/analyzer/language/python/binary/testdata/symlink
@@ -1,1 +1,0 @@
-pkg/fanal/analyzer/language/python/binary/testdata/foo
