+++ pkg/fanal/types/const.go
@@ -43,40 +43,40 @@ const (
 
 // Programming language dependencies
 const (
-	Bundler       LangType = "bundler"
-	GemSpec       LangType = "gemspec"
-	Cargo         LangType = "cargo"
-	PhpGeneric    LangType = "php"
-	Composer      LangType = "composer"
-	Npm           LangType = "npm"
-	NuGet         LangType = "nuget"
-	DotNetCore    LangType = "dotnet-core"
-	PackagesProps LangType = "packages-props"
-	Pip           LangType = "pip"
-	Pipenv        LangType = "pipenv"
-	Poetry        LangType = "poetry"
-	CondaPkg      LangType = "conda-pkg"
-	CondaEnv      LangType = "conda-environment"
-	PythonPkg     LangType = "python-pkg"
-	PythonGeneric LangType = "python"
-	NodeJsGeneric LangType = "nodejs"
-	NodePkg       LangType = "node-pkg"
-	Yarn          LangType = "yarn"
-	Pnpm          LangType = "pnpm"
-	Jar           LangType = "jar"
-	Pom           LangType = "pom"
-	Gradle        LangType = "gradle"
-	GoBinary      LangType = "gobinary"
-	GoModule      LangType = "gomod"
-	JavaScript    LangType = "javascript"
-	RustBinary    LangType = "rustbinary"
-	Conan         LangType = "conan"
-	Cocoapods     LangType = "cocoapods"
-	Swift         LangType = "swift"
-	Pub           LangType = "pub"
-	Hex           LangType = "hex"
-	Bitnami       LangType = "bitnami"
-	Julia         LangType = "julia"
+	Bundler          LangType = "bundler"
+	GemSpec          LangType = "gemspec"
+	Cargo            LangType = "cargo"
+	PhpExecutable    LangType = "php"
+	Composer         LangType = "composer"
+	Npm              LangType = "npm"
+	NuGet            LangType = "nuget"
+	DotNetCore       LangType = "dotnet-core"
+	PackagesProps    LangType = "packages-props"
+	Pip              LangType = "pip"
+	Pipenv           LangType = "pipenv"
+	Poetry           LangType = "poetry"
+	CondaPkg         LangType = "conda-pkg"
+	CondaEnv         LangType = "conda-environment"
+	PythonPkg        LangType = "python-pkg"
+	PythonExecutable LangType = "python"
+	NodeJsExecutable LangType = "nodejs"
+	NodePkg          LangType = "node-pkg"
+	Yarn             LangType = "yarn"
+	Pnpm             LangType = "pnpm"
+	Jar              LangType = "jar"
+	Pom              LangType = "pom"
+	Gradle           LangType = "gradle"
+	GoBinary         LangType = "gobinary"
+	GoModule         LangType = "gomod"
+	JavaScript       LangType = "javascript"
+	RustBinary       LangType = "rustbinary"
+	Conan            LangType = "conan"
+	Cocoapods        LangType = "cocoapods"
+	Swift            LangType = "swift"
+	Pub              LangType = "pub"
+	Hex              LangType = "hex"
+	Bitnami          LangType = "bitnami"
+	Julia            LangType = "julia"
 
 	K8sUpstream LangType = "kubernetes"
 	EKS         LangType = "eks" // Amazon Elastic Kubernetes Service
